RavenDB-24887 - Address PR Comments
