+++ src/Raven.Server/Documents/Handlers/AI/Agents/AiAgentProcessorForDeleteAiAgent.cs
@@ -9,6 +9,7 @@ using Raven.Server.ServerWide.Commands.AI;
 using Raven.Server.ServerWide.Context;
 using Sparrow.Json;
 using Sparrow.Json.Parsing;
+using Sparrow.Logging;
 
 namespace Raven.Server.Documents.Handlers.AI.Agents;
 
@@ -26,7 +27,8 @@ internal class AiAgentProcessorForDeleteAiAgent<TRequestHandler, TOperationConte
         var identifier = RequestHandler.GetStringQueryString("agentId", required: true);
         var r = await ServerStore.SendToLeaderAsync(new DeleteAiAgentCommand(RequestHandler.DatabaseName, identifier, RequestHandler.GetRaftRequestIdFromQuery()), token.Token);
 
-        RequestHandler.LogAuditForDatabase("DELETE", $"AiAgentConfiguration '{identifier}'");
+        if (RavenLogManager.Instance.IsAuditEnabled)
+            RequestHandler.LogAuditForDatabase("DELETE", $"AiAgentConfiguration '{identifier}'");
 
         await RequestHandler.WaitForIndexNotificationAsync(r.Index);
